+++ src/entities/user.entity.ts
@@ -1,4 +1,3 @@
-import crypto from "crypto";
 import { BeforeInsert, Column, Entity, Index, OneToMany } from "typeorm";
 import bcrypt from "bcryptjs";
 import Model from "./base.entity";