+++ src/index.ts
@@ -1,8 +1,13 @@
 import config from "./config/config";
 import app from "./server";
+import { PostgresDataSource } from "./utils/db";
 
-app.listen(config.appPort, () => {
-  console.log(
-    `⚡️[server]: Server is running at http://localhost:${config.appPort}`,
-  );
-});
+PostgresDataSource.initialize()
+  .then(async () => {
+    app.listen(config.appPort, () => {
+      console.log(
+        `⚡️[server]: Server is running at http://localhost:${config.appPort}`,
+      );
+    });
+  })
+  .catch((error) => console.log(error));