+++ src/utils/db.ts
@@ -1,15 +1,21 @@
+import "reflect-metadata";
 import { DataSource } from "typeorm";
 import config from "../config/config";
 
-export const PostgresDataSource = new DataSource({
-  type: "postgres",
+const postgresConfig = {
   host: config.database.host,
   port: config.database.port,
   username: config.database.user,
   password: config.database.password,
   database: config.database.name,
+};
+
+export const PostgresDataSource = new DataSource({
+  ...postgresConfig,
+  type: "postgres",
   synchronize: false,
   logging: false,
-  entities: ["webserver/entities/**/*.entity.{.ts,.js}"],
-  migrations: ["webserver/migrations/**/*{.ts,.js}"],
+  entities: [__dirname + "/../**/*.entity.{js,ts}"],
+  // entities: ["dist/entities/**/*.entity.{.ts,.js}"],
+  migrations: ["dist/migrations/**/*{.ts,.js}"],
 });