+++ test/user.ts
@@ -1,0 +1,7 @@
+export const User = {
+  first_name: "Test",
+  last_name: "Test",
+  username: "ttTest",
+  email: "test@test.net",
+  password: "test",
+};